Tối ưu hóa các controller Tag, Post, Category.

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,9 +3,10 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreCategoryFormRequest;
-use App\Http\Requests\UpdateCategoryFormRequest;
+use App\Http\Requests\Admin\CreateCategoryRequest;
+use App\Http\Requests\Admin\EditCategoryRequest;
 use App\Models\Category;
+use App\Models\Tag;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -13,32 +14,25 @@ class CategoryController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
-    {
+    public function index() {
         $categories = Category::paginate(8);
-
-        $data = [
-            'categories' => $categories,
-        ];
-
-        return view('admin.categories.index', $data);
+        return view('admin.categories.index', [
+            'categories' => $categories
+        ]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
-    {
+    public function create() {
         return view('admin.categories.create');
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreCategoryFormRequest $request)
-    {
-        $data = $request->all();
-        Category::create($data);
+    public function store(CreateCategoryRequest $request) {
+        Category::create($request->validated());
         return redirect()->route('admin.categories.index')->with('success', 'Thêm mới danh mục thành công.');
     }
 
@@ -53,35 +47,28 @@ class CategoryController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
-    {
-        $category = Category::find($id);
-
-        $data = [
-            'category' => $category,
-        ];
-
-        return view('admin.categories.edit', $data);
+    public function edit(Category $category) {
+        return view('admin.categories.edit', [
+            'category' => $category
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateCategoryFormRequest $request, string $id)
-    {
-        $data = $request->all();
-        $category = Category::find($id);
-        $category->update($data);
+    public function update(EditCategoryRequest $request, Category $category) {
+        $category->update($request->validated());
         return redirect()->route('admin.categories.index')->with('success', 'Cập nhật danh mục thành công.');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
-    {
-        $category = Category::find($id);
+    public function destroy(Category $category) {
+        if($category->posts()->count() > 0){
+            return redirect()->route('admin.categories.index')->with('success', 'Danh mục này có bài viết, không thể xóa.');
+        }
         $category->delete();
-        return redirect()->route('admin.categories.index')->with('success', 'Xóa danh mục ' . $category->name . ' thành công.');
+        return redirect()->route('admin.categories.index')->with('success', 'Xóa danh mục thành công.');
     }
 }
